backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
google-generativeai
groq
aiosqlite
PyPDF2
python-docx
sqlalchemy
pymysql
pydantic
cryptography
PyJWT
passlib
bcrypt